requirements.txt
anthropic
python-dotenv
google-generativeai
requests
